move tests and add new test for vertical level stddev
